+++ terraform/infra/ingress/cert-manager.tf
@@ -3,7 +3,7 @@ resource "helm_release" "cert-manager" {
   repository = "https://charts.jetstack.io"
   chart      = "cert-manager"
   namespace  = "cert-manager"
-  version    = "v1.13.1"
+  version    = "v1.14.4"
   create_namespace = true
 
   set {